refactor: move types to separate files
