+++ src/types/response.ts
@@ -1,0 +1,7 @@
+import { FailKeys } from './alert';
+
+export interface ErrorResponse {
+  httpStatus: FailKeys;
+  message: string;
+  errorCode: string;
+}
