+++ src/utils/og.ts
@@ -4,14 +4,11 @@ type OpenGraphType = {
   templateTitle?: string;
   logo?: string;
 };
-// !STARTERCONF This OG is generated from https://github.com/theodorusclarence/og
-// Please clone them and self-host if your site is going to be visited by many people.
-// Then change the url and the default logo.
+
 export function openGraph({
   siteName,
   templateTitle,
   description,
-  // !STARTERCONF Or, you can use my server with your own logo.
   logo = 'https://og.next-template-pyodol.vercel.app/images/logo.jpg',
 }: OpenGraphType): string {
   const ogLogo = encodeURIComponent(logo);
